[setting] #26 modify package structure
