+++ MID_SKHU/src/main/java/com/skhu/mid_skhu/app/dto/auth/RefreshTokenParsingDto.java
@@ -1,0 +1,12 @@
+package com.skhu.mid_skhu.app.dto.auth;
+
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+
+@Data
+@Builder
+public class RefreshTokenParsingDto {
+
+    private Long studentId;
+}
