+++ MID_SKHU/src/main/java/com/skhu/mid_skhu/app/dto/auth/RenewAccessTokenDto.java
@@ -1,0 +1,11 @@
+package com.skhu.mid_skhu.app.dto.auth;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class RenewAccessTokenDto {
+
+    private String renewAccessToken;
+}
