+++ MID_SKHU/src/main/java/com/skhu/mid_skhu/app/dto/auth/requestDto/LoginRequestDto.java
@@ -1,0 +1,10 @@
+package com.skhu.mid_skhu.app.dto.auth.requestDto;
+
+import lombok.Data;
+
+@Data
+public class LoginRequestDto {
+
+    private String studentNo;
+    private String password;
+}
