+++ MID_SKHU/src/main/java/com/skhu/mid_skhu/app/dto/auth/requestDto/SignUpRequestDto.java
@@ -1,0 +1,13 @@
+package com.skhu.mid_skhu.app.dto.auth.requestDto;
+
+import lombok.Data;
+
+@Data
+public class SignUpRequestDto {
+
+    private String studentNo;
+    private String password;
+    private String name;
+    private String phoneNumber;
+    private String roleType;
+}
