+++ MID_SKHU/src/main/java/com/skhu/mid_skhu/app/dto/auth/requestDto/StudentNoDuplicateInspectionRequestDto.java
@@ -1,0 +1,9 @@
+package com.skhu.mid_skhu.app.dto.auth.requestDto;
+
+import lombok.Data;
+
+@Data
+public class StudentNoDuplicateInspectionRequestDto {
+
+    private String studentNo;
+}
