+++ MID_SKHU/src/main/java/com/skhu/mid_skhu/app/dto/interest/requestDto/InterestRegisterRequestDto.java
@@ -1,0 +1,11 @@
+package com.skhu.mid_skhu.app.dto.interest.requestDto;
+
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+public class InterestRegisterRequestDto {
+
+    private List<String> interestCategoryList;
+}
