+++ MID_SKHU/src/main/java/com/skhu/mid_skhu/app/dto/interest/responseDto/InterestRegisterResponseDto.java
@@ -1,0 +1,13 @@
+package com.skhu.mid_skhu.app.dto.interest.responseDto;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@Builder
+public class InterestRegisterResponseDto {
+
+    private List<String> interestCategoryList;
+}
