--- MID_SKHU/src/main/java/com/skhu/mid_skhu/dto/auth/RefreshTokenParsingDto.java
@@ -1,12 +1,0 @@
-package com.skhu.mid_skhu.dto.auth;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-
-@Data
-@Builder
-public class RefreshTokenParsingDto {
-
-    private Long studentId;
-}
