--- MID_SKHU/src/main/java/com/skhu/mid_skhu/dto/auth/RenewAccessTokenDto.java
@@ -1,11 +1,0 @@
-package com.skhu.mid_skhu.dto.auth;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class RenewAccessTokenDto {
-
-    private String renewAccessToken;
-}
