--- MID_SKHU/src/main/java/com/skhu/mid_skhu/dto/auth/requestDto/LoginRequestDto.java
@@ -1,10 +1,0 @@
-package com.skhu.mid_skhu.dto.auth.requestDto;
-
-import lombok.Data;
-
-@Data
-public class LoginRequestDto {
-
-    private String studentNo;
-    private String password;
-}
