--- MID_SKHU/src/main/java/com/skhu/mid_skhu/dto/auth/requestDto/StudentNoDuplicateInspectionRequestDto.java
@@ -1,9 +1,0 @@
-package com.skhu.mid_skhu.dto.auth.requestDto;
-
-import lombok.Data;
-
-@Data
-public class StudentNoDuplicateInspectionRequestDto {
-
-    private String studentNo;
-}
