--- MID_SKHU/src/main/java/com/skhu/mid_skhu/dto/auth/responseDto/StudentNoDuplicateInspectionResponseDto.java
@@ -1,11 +1,0 @@
-package com.skhu.mid_skhu.dto.auth.responseDto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class StudentNoDuplicateInspectionResponseDto {
-
-    private Boolean studentNo;
-}
