--- MID_SKHU/src/main/java/com/skhu/mid_skhu/entity/student/StudentRefreshToken.java
@@ -1,40 +1,0 @@
-package com.skhu.mid_skhu.entity.student;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
-import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import net.minidev.json.annotate.JsonIgnore;
-
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "STUDENT_REFRESH_TOKEN")
-@Entity
-public class StudentRefreshToken {
-
-    @JsonIgnore
-    @Id
-    @Column(name = "REFRESH_TOKEN_SEQ")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long refreshTokenSeq;
-
-    @Column(name = "REFRESH_TOKEN", length = 256)
-    private String refreshToken;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "STUDENT_ID")
-    private Student student;
-}
